Database.Relational.Schema.IBMDB2

Description

This module implements queries to get table schema and table constraint informations from system catalog of IBM DB2.

Documentation

normalizeColumn :: String -> String

Normalize column name string to query DB2 system catalog

notNull :: Columns -> Bool

Not-null attribute information of column.

getType

Arguments

:: Map String TypeQ
Type mapping specified by user

-> Columns
Column info in system catalog

-> Maybe (String, TypeQ)
Result normalized name and mapped Haskell type

Get column normalized name and column Haskell type.

columnsQuerySQL :: Query (String, String) Columns

Phantom typed Query to get Columns from schema name and table name.

primaryKeyQuerySQL :: Query (String, String) String

Phantom typed Query to get primary key name from schema name and table name.