Database.Relational.Schema.PostgreSQL

Description

This module implements queries to get table schema and table constraint informations from system catalog of PostgreSQL.

Documentation

type Column = (PgAttribute, PgType)

Type to represent Column information.

normalizeColumn :: String -> String

Normalize column name string to query PostgreSQL system catalog.

notNull :: Column -> Bool

Not-null attribute information of column.

getType

Arguments

:: Map String TypeQ
Type mapping specified by user

-> Column
Column info in system catalog

-> Maybe (String, TypeQ)
Result normalized name and mapped Haskell type

Get column normalized name and column Haskell type.

columnQuerySQL :: Query (String, String) Column

Phantom typed Query to get Column from schema name and table name.

primaryKeyLengthQuerySQL :: Query (String, String) Int32

Phantom typed Query to get primary key length from schema name and table name.

primaryKeyQuerySQL :: Int32 -> Query (String, String) String

Phantom typed Query to get primary key name from schema name and table name.